moving catch_unwind to panic module
